+++ src/Symfony/Bridge/Twig/Extension/WorkflowExtension.php
@@ -21,6 +21,7 @@ use Twig\TwigFunction;
  * WorkflowExtension.
  *
  * @author Grégoire Pineau <lyrixx@lyrixx.info>
+ * @author Carlos Pereira De Amorim <carlos@shauri.fr>
  */
 final class WorkflowExtension extends AbstractExtension
 {