+++ src/Symfony/Component/Workflow/CHANGELOG.md
@@ -5,6 +5,8 @@ CHANGELOG
 -----
 
  * Added `Workflow::getEnabledTransition()` to easily retrieve a specific transition object
+ * Added context to the event dispatched
+ * Added default context to the Initial Marking
 
 5.1.0
 -----